requirements.txt
black
dj-config-url==0.1.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-extensions==4.1
django-graphql-jwt==0.4.0
graphene-django==3.2.3
gunicorn==25.1.0
pillow==12.1.1
psycopg2-binary==2.9.11
whitenoise==6.11.0
